Move datafusion proto definitions to datafusion-serialize
