+++ datafusion-proto/src/from_proto.rs
@@ -1,0 +1,1394 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+use crate::protobuf;
+use datafusion::arrow::datatypes::{
+    DataType, Field, IntervalUnit, Schema, TimeUnit, UnionMode,
+};
+use datafusion::error::DataFusionError;
+use datafusion::logical_plan::window_frames::{
+    WindowFrame, WindowFrameBound, WindowFrameUnits,
+};
+use datafusion::logical_plan::{
+    abs, atan, ceil, cos, digest, exp, floor, ln, log10, log2, round, signum, sin, sqrt,
+    tan, trunc, Column, DFField, DFSchema, DFSchemaRef, Expr, Operator,
+};
+use datafusion::physical_plan::aggregates::AggregateFunction;
+use datafusion::physical_plan::functions::BuiltinScalarFunction;
+use datafusion::physical_plan::window_functions::BuiltInWindowFunction;
+use datafusion::prelude::{
+    date_part, date_trunc, lower, ltrim, octet_length, rtrim, sha224, sha256, sha384,
+    sha512, trim, upper,
+};
+use datafusion::scalar::ScalarValue;
+use std::sync::Arc;
+
+#[derive(Debug)]
+pub enum Error {
+    General(String),
+    DataFusionError(DataFusionError),
+}
+
+impl std::fmt::Display for Error {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Self::General(desc) => write!(f, "General error: {}", desc),
+            Self::DataFusionError(desc) => {
+                write!(f, "DataFusion error: {:?}", desc)
+            }
+        }
+    }
+}
+
+impl std::error::Error for Error {}
+
+impl From<DataFusionError> for Error {
+    fn from(e: DataFusionError) -> Self {
+        Error::DataFusionError(e)
+    }
+}
+
+#[macro_export]
+macro_rules! convert_required {
+    ($PB:expr) => {{
+        if let Some(field) = $PB.as_ref() {
+            field.try_into()
+        } else {
+            Err(proto_error("Missing required field in protobuf"))
+        }
+    }};
+}
+
+#[macro_export]
+macro_rules! convert_box_required {
+    ($PB:expr) => {{
+        if let Some(field) = $PB.as_ref() {
+            field.as_ref().try_into()
+        } else {
+            Err(proto_error("Missing required field in protobuf"))
+        }
+    }};
+}
+
+impl From<protobuf::Column> for Column {
+    fn from(c: protobuf::Column) -> Self {
+        let protobuf::Column { relation, name } = c;
+
+        Self {
+            relation: relation.map(|r| r.relation),
+            name,
+        }
+    }
+}
+
+impl From<&protobuf::Column> for Column {
+    fn from(c: &protobuf::Column) -> Self {
+        c.clone().into()
+    }
+}
+
+impl TryFrom<&protobuf::DfSchema> for DFSchema {
+    type Error = Error;
+
+    fn try_from(df_schema: &protobuf::DfSchema) -> Result<Self, Self::Error> {
+        let fields = df_schema
+            .columns
+            .iter()
+            .map(|c| c.try_into())
+            .collect::<Result<Vec<DFField>, _>>()?;
+        Ok(DFSchema::new(fields)?)
+    }
+}
+
+impl TryFrom<protobuf::DfSchema> for DFSchemaRef {
+    type Error = Error;
+
+    fn try_from(df_schema: protobuf::DfSchema) -> Result<Self, Self::Error> {
+        let dfschema: DFSchema = (&df_schema).try_into()?;
+        Ok(Arc::new(dfschema))
+    }
+}
+
+impl TryFrom<&protobuf::DfField> for DFField {
+    type Error = Error;
+
+    fn try_from(df_field: &protobuf::DfField) -> Result<Self, Self::Error> {
+        let field: Field = convert_required!(df_field.field)?;
+
+        Ok(match &df_field.qualifier {
+            Some(q) => DFField::from_qualified(&q.relation, field),
+            None => DFField::from(field),
+        })
+    }
+}
+
+impl From<protobuf::WindowFrameUnits> for WindowFrameUnits {
+    fn from(units: protobuf::WindowFrameUnits) -> Self {
+        match units {
+            protobuf::WindowFrameUnits::Rows => Self::Rows,
+            protobuf::WindowFrameUnits::Range => Self::Range,
+            protobuf::WindowFrameUnits::Groups => Self::Groups,
+        }
+    }
+}
+
+impl From<protobuf::PrimitiveScalarType> for DataType {
+    fn from(scalar: protobuf::PrimitiveScalarType) -> Self {
+        match scalar {
+            protobuf::PrimitiveScalarType::Bool => DataType::Boolean,
+            protobuf::PrimitiveScalarType::Uint8 => DataType::UInt8,
+            protobuf::PrimitiveScalarType::Int8 => DataType::Int8,
+            protobuf::PrimitiveScalarType::Uint16 => DataType::UInt16,
+            protobuf::PrimitiveScalarType::Int16 => DataType::Int16,
+            protobuf::PrimitiveScalarType::Uint32 => DataType::UInt32,
+            protobuf::PrimitiveScalarType::Int32 => DataType::Int32,
+            protobuf::PrimitiveScalarType::Uint64 => DataType::UInt64,
+            protobuf::PrimitiveScalarType::Int64 => DataType::Int64,
+            protobuf::PrimitiveScalarType::Float32 => DataType::Float32,
+            protobuf::PrimitiveScalarType::Float64 => DataType::Float64,
+            protobuf::PrimitiveScalarType::Utf8 => DataType::Utf8,
+            protobuf::PrimitiveScalarType::LargeUtf8 => DataType::LargeUtf8,
+            protobuf::PrimitiveScalarType::Date32 => DataType::Date32,
+            protobuf::PrimitiveScalarType::TimeMicrosecond => {
+                DataType::Time64(TimeUnit::Microsecond)
+            }
+            protobuf::PrimitiveScalarType::TimeNanosecond => {
+                DataType::Time64(TimeUnit::Nanosecond)
+            }
+            protobuf::PrimitiveScalarType::Null => DataType::Null,
+            protobuf::PrimitiveScalarType::Decimal128 => DataType::Decimal(0, 0),
+            protobuf::PrimitiveScalarType::Date64 => DataType::Date64,
+            protobuf::PrimitiveScalarType::TimeSecond => {
+                DataType::Timestamp(TimeUnit::Second, None)
+            }
+            protobuf::PrimitiveScalarType::TimeMillisecond => {
+                DataType::Timestamp(TimeUnit::Millisecond, None)
+            }
+            protobuf::PrimitiveScalarType::IntervalYearmonth => {
+                DataType::Interval(IntervalUnit::YearMonth)
+            }
+            protobuf::PrimitiveScalarType::IntervalDaytime => {
+                DataType::Interval(IntervalUnit::DayTime)
+            }
+        }
+    }
+}
+
+impl TryFrom<&protobuf::ArrowType> for DataType {
+    type Error = Error;
+    fn try_from(arrow_type: &protobuf::ArrowType) -> Result<Self, Self::Error> {
+        let pb_arrow_type = arrow_type.arrow_type_enum.as_ref().ok_or_else(|| {
+            proto_error(
+                "Protobuf deserialization error: ArrowType missing required field 'data_type'",
+            )
+        })?;
+        pb_arrow_type.try_into()
+    }
+}
+
+impl TryFrom<&protobuf::arrow_type::ArrowTypeEnum> for DataType {
+    type Error = Error;
+    fn try_from(
+        arrow_type_enum: &protobuf::arrow_type::ArrowTypeEnum,
+    ) -> Result<Self, Self::Error> {
+        use protobuf::arrow_type;
+        Ok(match arrow_type_enum {
+            arrow_type::ArrowTypeEnum::None(_) => DataType::Null,
+            arrow_type::ArrowTypeEnum::Bool(_) => DataType::Boolean,
+            arrow_type::ArrowTypeEnum::Uint8(_) => DataType::UInt8,
+            arrow_type::ArrowTypeEnum::Int8(_) => DataType::Int8,
+            arrow_type::ArrowTypeEnum::Uint16(_) => DataType::UInt16,
+            arrow_type::ArrowTypeEnum::Int16(_) => DataType::Int16,
+            arrow_type::ArrowTypeEnum::Uint32(_) => DataType::UInt32,
+            arrow_type::ArrowTypeEnum::Int32(_) => DataType::Int32,
+            arrow_type::ArrowTypeEnum::Uint64(_) => DataType::UInt64,
+            arrow_type::ArrowTypeEnum::Int64(_) => DataType::Int64,
+            arrow_type::ArrowTypeEnum::Float16(_) => DataType::Float16,
+            arrow_type::ArrowTypeEnum::Float32(_) => DataType::Float32,
+            arrow_type::ArrowTypeEnum::Float64(_) => DataType::Float64,
+            arrow_type::ArrowTypeEnum::Utf8(_) => DataType::Utf8,
+            arrow_type::ArrowTypeEnum::LargeUtf8(_) => DataType::LargeUtf8,
+            arrow_type::ArrowTypeEnum::Binary(_) => DataType::Binary,
+            arrow_type::ArrowTypeEnum::FixedSizeBinary(size) => {
+                DataType::FixedSizeBinary(*size)
+            }
+            arrow_type::ArrowTypeEnum::LargeBinary(_) => DataType::LargeBinary,
+            arrow_type::ArrowTypeEnum::Date32(_) => DataType::Date32,
+            arrow_type::ArrowTypeEnum::Date64(_) => DataType::Date64,
+            arrow_type::ArrowTypeEnum::Duration(time_unit) => {
+                DataType::Duration(protobuf::TimeUnit::from_i32_to_arrow(*time_unit)?)
+            }
+            arrow_type::ArrowTypeEnum::Timestamp(protobuf::Timestamp {
+                time_unit,
+                timezone,
+            }) => DataType::Timestamp(
+                protobuf::TimeUnit::from_i32_to_arrow(*time_unit)?,
+                match timezone.len() {
+                    0 => None,
+                    _ => Some(timezone.to_owned()),
+                },
+            ),
+            arrow_type::ArrowTypeEnum::Time32(time_unit) => {
+                DataType::Time32(protobuf::TimeUnit::from_i32_to_arrow(*time_unit)?)
+            }
+            arrow_type::ArrowTypeEnum::Time64(time_unit) => {
+                DataType::Time64(protobuf::TimeUnit::from_i32_to_arrow(*time_unit)?)
+            }
+            arrow_type::ArrowTypeEnum::Interval(interval_unit) => DataType::Interval(
+                protobuf::IntervalUnit::from_i32_to_arrow(*interval_unit)?,
+            ),
+            arrow_type::ArrowTypeEnum::Decimal(protobuf::Decimal {
+                whole,
+                fractional,
+            }) => DataType::Decimal(*whole as usize, *fractional as usize),
+            arrow_type::ArrowTypeEnum::List(list) => {
+                let list_type: &protobuf::Field = list
+                    .as_ref()
+                    .field_type
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: List message missing required field 'field_type'"))?
+                    .as_ref();
+                DataType::List(Box::new(list_type.try_into()?))
+            }
+            arrow_type::ArrowTypeEnum::LargeList(list) => {
+                let list_type: &protobuf::Field = list
+                    .as_ref()
+                    .field_type
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: List message missing required field 'field_type'"))?
+                    .as_ref();
+                DataType::LargeList(Box::new(list_type.try_into()?))
+            }
+            arrow_type::ArrowTypeEnum::FixedSizeList(list) => {
+                let list_type: &protobuf::Field = list
+                    .as_ref()
+                    .field_type
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: List message missing required field 'field_type'"))?
+                    .as_ref();
+                let list_size = list.list_size;
+                DataType::FixedSizeList(Box::new(list_type.try_into()?), list_size)
+            }
+            arrow_type::ArrowTypeEnum::Struct(strct) => DataType::Struct(
+                strct
+                    .sub_field_types
+                    .iter()
+                    .map(|field| field.try_into())
+                    .collect::<Result<Vec<_>, _>>()?,
+            ),
+            arrow_type::ArrowTypeEnum::Union(union) => {
+                let union_mode = protobuf::UnionMode::from_i32(union.union_mode)
+                    .ok_or_else(|| {
+                        proto_error(
+                            "Protobuf deserialization error: Unknown union mode type",
+                        )
+                    })?;
+                let union_mode = match union_mode {
+                    protobuf::UnionMode::Dense => UnionMode::Dense,
+                    protobuf::UnionMode::Sparse => UnionMode::Sparse,
+                };
+                let union_types = union
+                    .union_types
+                    .iter()
+                    .map(|field| field.try_into())
+                    .collect::<Result<Vec<_>, _>>()?;
+                DataType::Union(union_types, union_mode)
+            }
+            arrow_type::ArrowTypeEnum::Dictionary(dict) => {
+                let pb_key_datatype = dict
+                    .as_ref()
+                    .key
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: Dictionary message missing required field 'key'"))?;
+                let pb_value_datatype = dict
+                    .as_ref()
+                    .value
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: Dictionary message missing required field 'key'"))?;
+                let key_datatype: DataType = pb_key_datatype.as_ref().try_into()?;
+                let value_datatype: DataType = pb_value_datatype.as_ref().try_into()?;
+                DataType::Dictionary(Box::new(key_datatype), Box::new(value_datatype))
+            }
+        })
+    }
+}
+
+impl TryFrom<&protobuf::Field> for Field {
+    type Error = Error;
+    fn try_from(field: &protobuf::Field) -> Result<Self, Self::Error> {
+        let pb_datatype = field.arrow_type.as_ref().ok_or_else(|| {
+            proto_error(
+                "Protobuf deserialization error: Field message missing required field 'arrow_type'",
+            )
+        })?;
+
+        Ok(Self::new(
+            field.name.as_str(),
+            pb_datatype.as_ref().try_into()?,
+            field.nullable,
+        ))
+    }
+}
+
+impl From<&protobuf::ScalarFunction> for BuiltinScalarFunction {
+    fn from(f: &protobuf::ScalarFunction) -> Self {
+        use protobuf::ScalarFunction;
+        match f {
+            ScalarFunction::Sqrt => Self::Sqrt,
+            ScalarFunction::Sin => Self::Sin,
+            ScalarFunction::Cos => Self::Cos,
+            ScalarFunction::Tan => Self::Tan,
+            ScalarFunction::Asin => Self::Asin,
+            ScalarFunction::Acos => Self::Acos,
+            ScalarFunction::Atan => Self::Atan,
+            ScalarFunction::Exp => Self::Exp,
+            ScalarFunction::Log => Self::Log,
+            ScalarFunction::Log2 => Self::Log2,
+            ScalarFunction::Log10 => Self::Log10,
+            ScalarFunction::Floor => Self::Floor,
+            ScalarFunction::Ceil => Self::Ceil,
+            ScalarFunction::Round => Self::Round,
+            ScalarFunction::Trunc => Self::Trunc,
+            ScalarFunction::Abs => Self::Abs,
+            ScalarFunction::Signum => Self::Signum,
+            ScalarFunction::Octetlength => Self::OctetLength,
+            ScalarFunction::Concat => Self::Concat,
+            ScalarFunction::Lower => Self::Lower,
+            ScalarFunction::Upper => Self::Upper,
+            ScalarFunction::Trim => Self::Trim,
+            ScalarFunction::Ltrim => Self::Ltrim,
+            ScalarFunction::Rtrim => Self::Rtrim,
+            ScalarFunction::Totimestamp => Self::ToTimestamp,
+            ScalarFunction::Array => Self::Array,
+            ScalarFunction::Nullif => Self::NullIf,
+            ScalarFunction::Datepart => Self::DatePart,
+            ScalarFunction::Datetrunc => Self::DateTrunc,
+            ScalarFunction::Md5 => Self::MD5,
+            ScalarFunction::Sha224 => Self::SHA224,
+            ScalarFunction::Sha256 => Self::SHA256,
+            ScalarFunction::Sha384 => Self::SHA384,
+            ScalarFunction::Sha512 => Self::SHA512,
+            ScalarFunction::Digest => Self::Digest,
+            ScalarFunction::Ln => Self::Ln,
+            ScalarFunction::Totimestampmillis => Self::ToTimestampMillis,
+        }
+    }
+}
+
+impl From<protobuf::AggregateFunction> for AggregateFunction {
+    fn from(agg_fun: protobuf::AggregateFunction) -> Self {
+        match agg_fun {
+            protobuf::AggregateFunction::Min => Self::Min,
+            protobuf::AggregateFunction::Max => Self::Max,
+            protobuf::AggregateFunction::Sum => Self::Sum,
+            protobuf::AggregateFunction::Avg => Self::Avg,
+            protobuf::AggregateFunction::Count => Self::Count,
+            protobuf::AggregateFunction::ApproxDistinct => Self::ApproxDistinct,
+            protobuf::AggregateFunction::ArrayAgg => Self::ArrayAgg,
+            protobuf::AggregateFunction::Variance => Self::Variance,
+            protobuf::AggregateFunction::VariancePop => Self::VariancePop,
+            protobuf::AggregateFunction::Covariance => Self::Covariance,
+            protobuf::AggregateFunction::CovariancePop => Self::CovariancePop,
+            protobuf::AggregateFunction::Stddev => Self::Stddev,
+            protobuf::AggregateFunction::StddevPop => Self::StddevPop,
+            protobuf::AggregateFunction::Correlation => Self::Correlation,
+            protobuf::AggregateFunction::ApproxPercentileCont => {
+                Self::ApproxPercentileCont
+            }
+            protobuf::AggregateFunction::ApproxMedian => Self::ApproxMedian,
+        }
+    }
+}
+
+impl From<protobuf::BuiltInWindowFunction> for BuiltInWindowFunction {
+    fn from(built_in_function: protobuf::BuiltInWindowFunction) -> Self {
+        match built_in_function {
+            protobuf::BuiltInWindowFunction::RowNumber => Self::RowNumber,
+            protobuf::BuiltInWindowFunction::Rank => Self::Rank,
+            protobuf::BuiltInWindowFunction::PercentRank => Self::PercentRank,
+            protobuf::BuiltInWindowFunction::DenseRank => Self::DenseRank,
+            protobuf::BuiltInWindowFunction::Lag => Self::Lag,
+            protobuf::BuiltInWindowFunction::Lead => Self::Lead,
+            protobuf::BuiltInWindowFunction::FirstValue => Self::FirstValue,
+            protobuf::BuiltInWindowFunction::CumeDist => Self::CumeDist,
+            protobuf::BuiltInWindowFunction::Ntile => Self::Ntile,
+            protobuf::BuiltInWindowFunction::NthValue => Self::NthValue,
+            protobuf::BuiltInWindowFunction::LastValue => Self::LastValue,
+        }
+    }
+}
+
+impl TryFrom<&protobuf::scalar_type::Datatype> for DataType {
+    type Error = Error;
+
+    fn try_from(
+        scalar_type: &protobuf::scalar_type::Datatype,
+    ) -> Result<Self, Self::Error> {
+        use protobuf::scalar_type::Datatype;
+
+        Ok(match scalar_type {
+            Datatype::Scalar(scalar_type) => {
+                let pb_scalar_enum = protobuf::PrimitiveScalarType::from_i32(*scalar_type).ok_or_else(|| {
+                    proto_error(format!(
+                        "Protobuf deserialization error, scalar_type::Datatype missing was provided invalid enum variant: {}",
+                        *scalar_type
+                    ))
+                })?;
+                pb_scalar_enum.into()
+            }
+            Datatype::List(protobuf::ScalarListType {
+                deepest_type,
+                field_names,
+            }) => {
+                if field_names.is_empty() {
+                    return Err(proto_error(
+                        "Protobuf deserialization error: found no field names in ScalarListType message which requires at least one",
+                    ));
+                }
+                let pb_scalar_type = protobuf::PrimitiveScalarType::from_i32(
+                    *deepest_type,
+                )
+                .ok_or_else(|| {
+                    proto_error(format!(
+                        "Protobuf deserialization error: invalid i32 for scalar enum: {}",
+                        *deepest_type
+                    ))
+                })?;
+                //Because length is checked above it is safe to unwrap .last()
+                let mut scalar_type = DataType::List(Box::new(Field::new(
+                    field_names.last().unwrap().as_str(),
+                    pb_scalar_type.into(),
+                    true,
+                )));
+                //Iterate over field names in reverse order except for the last item in the vector
+                for name in field_names.iter().rev().skip(1) {
+                    let new_datatype = DataType::List(Box::new(Field::new(
+                        name.as_str(),
+                        scalar_type,
+                        true,
+                    )));
+                    scalar_type = new_datatype;
+                }
+                scalar_type
+            }
+        })
+    }
+}
+
+impl TryFrom<&protobuf::scalar_value::Value> for ScalarValue {
+    type Error = Error;
+
+    fn try_from(scalar: &protobuf::scalar_value::Value) -> Result<Self, Self::Error> {
+        use protobuf::{scalar_value::Value, PrimitiveScalarType};
+
+        let scalar = match scalar {
+            Value::BoolValue(v) => ScalarValue::Boolean(Some(*v)),
+            Value::Utf8Value(v) => ScalarValue::Utf8(Some(v.to_owned())),
+            Value::LargeUtf8Value(v) => ScalarValue::LargeUtf8(Some(v.to_owned())),
+            Value::Int8Value(v) => ScalarValue::Int8(Some(*v as i8)),
+            Value::Int16Value(v) => ScalarValue::Int16(Some(*v as i16)),
+            Value::Int32Value(v) => ScalarValue::Int32(Some(*v)),
+            Value::Int64Value(v) => ScalarValue::Int64(Some(*v)),
+            Value::Uint8Value(v) => ScalarValue::UInt8(Some(*v as u8)),
+            Value::Uint16Value(v) => ScalarValue::UInt16(Some(*v as u16)),
+            Value::Uint32Value(v) => ScalarValue::UInt32(Some(*v)),
+            Value::Uint64Value(v) => ScalarValue::UInt64(Some(*v)),
+            Value::Float32Value(v) => ScalarValue::Float32(Some(*v)),
+            Value::Float64Value(v) => ScalarValue::Float64(Some(*v)),
+            Value::Date32Value(v) => ScalarValue::Date32(Some(*v)),
+            Value::TimeMicrosecondValue(v) => {
+                ScalarValue::TimestampMicrosecond(Some(*v), None)
+            }
+            Value::TimeNanosecondValue(v) => {
+                ScalarValue::TimestampNanosecond(Some(*v), None)
+            }
+            Value::ListValue(v) => v.try_into()?,
+            Value::NullListValue(v) => ScalarValue::List(None, Box::new(v.try_into()?)),
+            Value::NullValue(null_enum) => {
+                let primitive = PrimitiveScalarType::from_i32(*null_enum)
+                    .ok_or_else(|| proto_error("Invalid scalar type"))?;
+                (&primitive).try_into()?
+            }
+            Value::Decimal128Value(val) => {
+                let array = vec_to_array(val.value.clone());
+                ScalarValue::Decimal128(
+                    Some(i128::from_be_bytes(array)),
+                    val.p as usize,
+                    val.s as usize,
+                )
+            }
+            Value::Date64Value(v) => ScalarValue::Date64(Some(*v)),
+            Value::TimeSecondValue(v) => ScalarValue::TimestampSecond(Some(*v), None),
+            Value::TimeMillisecondValue(v) => {
+                ScalarValue::TimestampMillisecond(Some(*v), None)
+            }
+            Value::IntervalYearmonthValue(v) => ScalarValue::IntervalYearMonth(Some(*v)),
+            Value::IntervalDaytimeValue(v) => ScalarValue::IntervalDayTime(Some(*v)),
+        };
+        Ok(scalar)
+    }
+}
+
+impl TryFrom<&protobuf::ScalarListValue> for ScalarValue {
+    type Error = Error;
+
+    fn try_from(
+        scalar_list_value: &protobuf::ScalarListValue,
+    ) -> Result<Self, Self::Error> {
+        use protobuf::{scalar_type::Datatype, PrimitiveScalarType};
+
+        let protobuf::ScalarListValue { datatype, values } = scalar_list_value;
+        let pb_scalar_type = datatype
+            .as_ref()
+            .ok_or_else(|| proto_error("Protobuf deserialization error: ScalarListValue messsage missing required field 'datatype'"))?;
+        let scalar_type = pb_scalar_type
+            .datatype
+            .as_ref()
+            .ok_or_else(|| proto_error("Protobuf deserialization error: ScalarListValue.Datatype messsage missing required field 'datatype'"))?;
+        let scalar_values = match scalar_type {
+            Datatype::Scalar(scalar_type_i32) => {
+                let leaf_scalar_type =
+                    protobuf::PrimitiveScalarType::from_i32(*scalar_type_i32)
+                        .ok_or_else(|| {
+                            proto_error("Error converting i32 to basic scalar type")
+                        })?;
+                let typechecked_values: Vec<datafusion::scalar::ScalarValue> = values
+                    .iter()
+                    .map(|protobuf::ScalarValue { value: opt_value }| {
+                        let value = opt_value.as_ref().ok_or_else(|| {
+                            proto_error(
+                                "Protobuf deserialization error: missing required field 'value'",
+                            )
+                        })?;
+                        typechecked_scalar_value_conversion(value, leaf_scalar_type)
+                    })
+                    .collect::<Result<Vec<_>, _>>()?;
+                ScalarValue::List(
+                    Some(Box::new(typechecked_values)),
+                    Box::new(leaf_scalar_type.into()),
+                )
+            }
+            Datatype::List(list_type) => {
+                let protobuf::ScalarListType {
+                    deepest_type,
+                    field_names,
+                } = &list_type;
+                let leaf_type =
+                    PrimitiveScalarType::from_i32(*deepest_type).ok_or_else(|| {
+                        proto_error("Error converting i32 to basic scalar type")
+                    })?;
+                let depth = field_names.len();
+
+                let typechecked_values: Vec<ScalarValue> = if depth == 0 {
+                    return Err(proto_error(
+                        "Protobuf deserialization error, ScalarListType had no field names, requires at least one",
+                    ));
+                } else if depth == 1 {
+                    values
+                        .iter()
+                        .map(|protobuf::ScalarValue { value: opt_value }| {
+                            let value = opt_value
+                                .as_ref()
+                                .ok_or_else(|| proto_error("Protobuf deserialization error: missing required field 'value'"))?;
+                            typechecked_scalar_value_conversion(value, leaf_type)
+                        })
+                        .collect::<Result<Vec<_>, _>>()?
+                } else {
+                    values
+                        .iter()
+                        .map(|protobuf::ScalarValue { value: opt_value }| {
+                            let value = opt_value
+                                .as_ref()
+                                .ok_or_else(|| proto_error("Protobuf deserialization error: missing required field 'value'"))?;
+                            value.try_into()
+                        })
+                        .collect::<Result<Vec<_>, _>>()?
+                };
+                ScalarValue::List(
+                    match typechecked_values.len() {
+                        0 => None,
+                        _ => Some(Box::new(typechecked_values)),
+                    },
+                    Box::new((list_type).try_into()?),
+                )
+            }
+        };
+        Ok(scalar_values)
+    }
+}
+
+impl TryFrom<&protobuf::ScalarType> for DataType {
+    type Error = Error;
+    fn try_from(scalar: &protobuf::ScalarType) -> Result<Self, Self::Error> {
+        let pb_scalartype = scalar.datatype.as_ref().ok_or_else(|| {
+            proto_error("ScalarType message missing required field 'datatype'")
+        })?;
+        pb_scalartype.try_into()
+    }
+}
+
+impl TryFrom<&protobuf::Schema> for Schema {
+    type Error = Error;
+
+    fn try_from(schema: &protobuf::Schema) -> Result<Self, Self::Error> {
+        let fields = schema
+            .columns
+            .iter()
+            .map(|c| {
+                let pb_arrow_type_res = c
+                    .arrow_type
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: Field message was missing required field 'arrow_type'"));
+                let pb_arrow_type: &protobuf::ArrowType = match pb_arrow_type_res {
+                    Ok(res) => res,
+                    Err(e) => return Err(e),
+                };
+                Ok(Field::new(&c.name, pb_arrow_type.try_into()?, c.nullable))
+            })
+            .collect::<Result<Vec<_>, _>>()?;
+        Ok(Self::new(fields))
+    }
+}
+
+impl TryFrom<&protobuf::PrimitiveScalarType> for ScalarValue {
+    type Error = Error;
+
+    fn try_from(scalar: &protobuf::PrimitiveScalarType) -> Result<Self, Self::Error> {
+        use protobuf::PrimitiveScalarType;
+
+        Ok(match scalar {
+            PrimitiveScalarType::Null => {
+                return Err(proto_error("Untyped null is an invalid scalar value"))
+            }
+            PrimitiveScalarType::Bool => Self::Boolean(None),
+            PrimitiveScalarType::Uint8 => Self::UInt8(None),
+            PrimitiveScalarType::Int8 => Self::Int8(None),
+            PrimitiveScalarType::Uint16 => Self::UInt16(None),
+            PrimitiveScalarType::Int16 => Self::Int16(None),
+            PrimitiveScalarType::Uint32 => Self::UInt32(None),
+            PrimitiveScalarType::Int32 => Self::Int32(None),
+            PrimitiveScalarType::Uint64 => Self::UInt64(None),
+            PrimitiveScalarType::Int64 => Self::Int64(None),
+            PrimitiveScalarType::Float32 => Self::Float32(None),
+            PrimitiveScalarType::Float64 => Self::Float64(None),
+            PrimitiveScalarType::Utf8 => Self::Utf8(None),
+            PrimitiveScalarType::LargeUtf8 => Self::LargeUtf8(None),
+            PrimitiveScalarType::Date32 => Self::Date32(None),
+            PrimitiveScalarType::TimeMicrosecond => {
+                Self::TimestampMicrosecond(None, None)
+            }
+            PrimitiveScalarType::TimeNanosecond => Self::TimestampNanosecond(None, None),
+            PrimitiveScalarType::Decimal128 => Self::Decimal128(None, 0, 0),
+            PrimitiveScalarType::Date64 => Self::Date64(None),
+            PrimitiveScalarType::TimeSecond => Self::TimestampSecond(None, None),
+            PrimitiveScalarType::TimeMillisecond => {
+                Self::TimestampMillisecond(None, None)
+            }
+            PrimitiveScalarType::IntervalYearmonth => Self::IntervalYearMonth(None),
+            PrimitiveScalarType::IntervalDaytime => Self::IntervalDayTime(None),
+        })
+    }
+}
+
+impl TryFrom<&protobuf::ScalarListType> for DataType {
+    type Error = Error;
+    fn try_from(scalar: &protobuf::ScalarListType) -> Result<Self, Self::Error> {
+        use protobuf::PrimitiveScalarType;
+
+        let protobuf::ScalarListType {
+            deepest_type,
+            field_names,
+        } = scalar;
+
+        let depth = field_names.len();
+        if depth == 0 {
+            return Err(proto_error(
+                "Protobuf deserialization error: Found a ScalarListType message with no field names, at least one is required",
+            ));
+        }
+
+        let mut curr_type = Self::List(Box::new(Field::new(
+            //Since checked vector is not empty above this is safe to unwrap
+            field_names.last().unwrap(),
+            PrimitiveScalarType::from_i32(*deepest_type)
+                .ok_or_else(|| {
+                    proto_error("Could not convert to datafusion scalar type")
+                })?
+                .into(),
+            true,
+        )));
+        //Iterates over field names in reverse order except for the last item in the vector
+        for name in field_names.iter().rev().skip(1) {
+            let temp_curr_type = Self::List(Box::new(Field::new(name, curr_type, true)));
+            curr_type = temp_curr_type;
+        }
+        Ok(curr_type)
+    }
+}
+
+impl TryFrom<&protobuf::ScalarValue> for ScalarValue {
+    type Error = Error;
+
+    fn try_from(scalar: &protobuf::ScalarValue) -> Result<Self, Self::Error> {
+        use protobuf::scalar_value::Value;
+
+        let value = scalar.value.as_ref().ok_or_else(|| {
+            proto_error("Protobuf deserialization error: missing required field 'value'")
+        })?;
+        Ok(match value {
+            Value::BoolValue(v) => Self::Boolean(Some(*v)),
+            Value::Utf8Value(v) => Self::Utf8(Some(v.to_owned())),
+            Value::LargeUtf8Value(v) => Self::LargeUtf8(Some(v.to_owned())),
+            Value::Int8Value(v) => Self::Int8(Some(*v as i8)),
+            Value::Int16Value(v) => Self::Int16(Some(*v as i16)),
+            Value::Int32Value(v) => Self::Int32(Some(*v)),
+            Value::Int64Value(v) => Self::Int64(Some(*v)),
+            Value::Uint8Value(v) => Self::UInt8(Some(*v as u8)),
+            Value::Uint16Value(v) => Self::UInt16(Some(*v as u16)),
+            Value::Uint32Value(v) => Self::UInt32(Some(*v)),
+            Value::Uint64Value(v) => Self::UInt64(Some(*v)),
+            Value::Float32Value(v) => Self::Float32(Some(*v)),
+            Value::Float64Value(v) => Self::Float64(Some(*v)),
+            Value::Date32Value(v) => Self::Date32(Some(*v)),
+            Value::TimeMicrosecondValue(v) => Self::TimestampMicrosecond(Some(*v), None),
+            Value::TimeNanosecondValue(v) => Self::TimestampNanosecond(Some(*v), None),
+            Value::ListValue(scalar_list) => {
+                let protobuf::ScalarListValue {
+                    values,
+                    datatype: opt_scalar_type,
+                } = &scalar_list;
+                let pb_scalar_type = opt_scalar_type
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization err: ScalaListValue missing required field 'datatype'"))?;
+                let typechecked_values: Vec<ScalarValue> = values
+                    .iter()
+                    .map(|val| val.try_into())
+                    .collect::<Result<Vec<_>, _>>()?;
+                let scalar_type: DataType = pb_scalar_type.try_into()?;
+                let scalar_type = Box::new(scalar_type);
+                Self::List(Some(Box::new(typechecked_values)), scalar_type)
+            }
+            Value::NullListValue(v) => {
+                let pb_datatype = v
+                    .datatype
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: NullListValue message missing required field 'datatyp'"))?;
+                let pb_datatype = Box::new(pb_datatype.try_into()?);
+                Self::List(None, pb_datatype)
+            }
+            Value::NullValue(v) => {
+                let null_type_enum = protobuf::PrimitiveScalarType::from_i32(*v)
+                    .ok_or_else(|| proto_error("Protobuf deserialization error found invalid enum variant for DatafusionScalar"))?;
+                (&null_type_enum).try_into()?
+            }
+            Value::Decimal128Value(val) => {
+                let array = vec_to_array(val.value.clone());
+                Self::Decimal128(
+                    Some(i128::from_be_bytes(array)),
+                    val.p as usize,
+                    val.s as usize,
+                )
+            }
+            Value::Date64Value(v) => Self::Date64(Some(*v)),
+            Value::TimeSecondValue(v) => Self::TimestampSecond(Some(*v), None),
+            Value::TimeMillisecondValue(v) => Self::TimestampMillisecond(Some(*v), None),
+            Value::IntervalYearmonthValue(v) => Self::IntervalYearMonth(Some(*v)),
+            Value::IntervalDaytimeValue(v) => Self::IntervalDayTime(Some(*v)),
+        })
+    }
+}
+
+impl TryFrom<&protobuf::LogicalExprNode> for Expr {
+    type Error = Error;
+
+    fn try_from(expr: &protobuf::LogicalExprNode) -> Result<Self, Self::Error> {
+        use datafusion::physical_plan::window_functions;
+        use protobuf::{logical_expr_node::ExprType, window_expr_node, ScalarFunction};
+
+        let expr_type = expr
+            .expr_type
+            .as_ref()
+            .ok_or_else(|| proto_error("Unexpected empty logical expression"))?;
+        match expr_type {
+            ExprType::BinaryExpr(binary_expr) => Ok(Self::BinaryExpr {
+                left: Box::new(parse_required_expr(&binary_expr.l)?),
+                op: from_proto_binary_op(&binary_expr.op)?,
+                right: Box::new(parse_required_expr(&binary_expr.r)?),
+            }),
+            ExprType::Column(column) => Ok(Self::Column(column.into())),
+            ExprType::Literal(literal) => {
+                let scalar_value: ScalarValue = literal.try_into()?;
+                Ok(Self::Literal(scalar_value))
+            }
+            ExprType::WindowExpr(expr) => {
+                let window_function = expr
+                    .window_function
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Received empty window function"))?;
+                let partition_by = expr
+                    .partition_by
+                    .iter()
+                    .map(|e| e.try_into())
+                    .collect::<Result<Vec<_>, _>>()?;
+                let order_by = expr
+                    .order_by
+                    .iter()
+                    .map(|e| e.try_into())
+                    .collect::<Result<Vec<_>, _>>()?;
+                let window_frame = expr
+                    .window_frame
+                    .as_ref()
+                    .map::<Result<WindowFrame, _>, _>(|e| match e {
+                        window_expr_node::WindowFrame::Frame(frame) => {
+                            let window_frame: WindowFrame = frame.clone().try_into()?;
+                            if WindowFrameUnits::Range == window_frame.units
+                                && order_by.len() != 1
+                            {
+                                Err(proto_error("With window frame of type RANGE, the order by expression must be of length 1"))
+                            } else {
+                                Ok(window_frame)
+                            }
+                        }
+                    })
+                    .transpose()?;
+
+                match window_function {
+                    window_expr_node::WindowFunction::AggrFunction(i) => {
+                        let aggr_function = protobuf::AggregateFunction::from_i32(*i)
+                            .ok_or_else(|| {
+                                proto_error(format!(
+                                    "Received an unknown aggregate window function: {}",
+                                    i
+                                ))
+                            })?;
+
+                        Ok(Self::WindowFunction {
+                            fun: window_functions::WindowFunction::AggregateFunction(
+                                AggregateFunction::from(aggr_function),
+                            ),
+                            args: vec![parse_required_expr(&expr.expr)?],
+                            partition_by,
+                            order_by,
+                            window_frame,
+                        })
+                    }
+                    window_expr_node::WindowFunction::BuiltInFunction(i) => {
+                        let built_in_function =
+                            protobuf::BuiltInWindowFunction::from_i32(*i).ok_or_else(
+                                || {
+                                    proto_error(format!(
+                                        "Received an unknown built-in window function: {}",
+                                        i
+                                    ))
+                                },
+                            )?;
+
+                        Ok(Self::WindowFunction {
+                            fun: window_functions::WindowFunction::BuiltInWindowFunction(
+                                BuiltInWindowFunction::from(built_in_function),
+                            ),
+                            args: vec![parse_required_expr(&expr.expr)?],
+                            partition_by,
+                            order_by,
+                            window_frame,
+                        })
+                    }
+                }
+            }
+            ExprType::AggregateExpr(expr) => {
+                let aggr_function =
+                    protobuf::AggregateFunction::from_i32(expr.aggr_function)
+                        .ok_or_else(|| {
+                            proto_error(format!(
+                                "Received an unknown aggregate function: {}",
+                                expr.aggr_function
+                            ))
+                        })?;
+                let fun = AggregateFunction::from(aggr_function);
+
+                Ok(Self::AggregateFunction {
+                    fun,
+                    args: expr
+                        .expr
+                        .iter()
+                        .map(|e| e.try_into())
+                        .collect::<Result<Vec<_>, _>>()?,
+                    distinct: false, //TODO
+                })
+            }
+            ExprType::Alias(alias) => Ok(Self::Alias(
+                Box::new(parse_required_expr(&alias.expr)?),
+                alias.alias.clone(),
+            )),
+            ExprType::IsNullExpr(is_null) => {
+                Ok(Self::IsNull(Box::new(parse_required_expr(&is_null.expr)?)))
+            }
+            ExprType::IsNotNullExpr(is_not_null) => Ok(Self::IsNotNull(Box::new(
+                parse_required_expr(&is_not_null.expr)?,
+            ))),
+            ExprType::NotExpr(not) => {
+                Ok(Self::Not(Box::new(parse_required_expr(&not.expr)?)))
+            }
+            ExprType::Between(between) => Ok(Self::Between {
+                expr: Box::new(parse_required_expr(&between.expr)?),
+                negated: between.negated,
+                low: Box::new(parse_required_expr(&between.low)?),
+                high: Box::new(parse_required_expr(&between.high)?),
+            }),
+            ExprType::Case(case) => {
+                let when_then_expr = case
+                    .when_then_expr
+                    .iter()
+                    .map(|e| {
+                        Ok((
+                            Box::new(match &e.when_expr {
+                                Some(e) => e.try_into(),
+                                None => Err(proto_error("Missing required expression")),
+                            }?),
+                            Box::new(match &e.then_expr {
+                                Some(e) => e.try_into(),
+                                None => Err(proto_error("Missing required expression")),
+                            }?),
+                        ))
+                    })
+                    .collect::<Result<Vec<(Box<Expr>, Box<Expr>)>, Error>>()?;
+                Ok(Self::Case {
+                    expr: parse_optional_expr(&case.expr)?.map(Box::new),
+                    when_then_expr,
+                    else_expr: parse_optional_expr(&case.else_expr)?.map(Box::new),
+                })
+            }
+            ExprType::Cast(cast) => {
+                let expr = Box::new(parse_required_expr(&cast.expr)?);
+                let arrow_type: &protobuf::ArrowType = cast
+                    .arrow_type
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: CastNode message missing required field 'arrow_type'"))?;
+                let data_type = arrow_type.try_into()?;
+                Ok(Self::Cast { expr, data_type })
+            }
+            ExprType::TryCast(cast) => {
+                let expr = Box::new(parse_required_expr(&cast.expr)?);
+                let arrow_type: &protobuf::ArrowType = cast
+                    .arrow_type
+                    .as_ref()
+                    .ok_or_else(|| proto_error("Protobuf deserialization error: CastNode message missing required field 'arrow_type'"))?;
+                let data_type = arrow_type.try_into()?;
+                Ok(Self::TryCast { expr, data_type })
+            }
+            ExprType::Sort(sort) => Ok(Self::Sort {
+                expr: Box::new(parse_required_expr(&sort.expr)?),
+                asc: sort.asc,
+                nulls_first: sort.nulls_first,
+            }),
+            ExprType::Negative(negative) => Ok(Self::Negative(Box::new(
+                parse_required_expr(&negative.expr)?,
+            ))),
+            ExprType::InList(in_list) => Ok(Self::InList {
+                expr: Box::new(parse_required_expr(&in_list.expr)?),
+                list: in_list
+                    .list
+                    .iter()
+                    .map(|expr| expr.try_into())
+                    .collect::<Result<Vec<_>, _>>()?,
+                negated: in_list.negated,
+            }),
+            ExprType::Wildcard(_) => Ok(Self::Wildcard),
+            ExprType::ScalarFunction(expr) => {
+                let scalar_function = protobuf::ScalarFunction::from_i32(expr.fun)
+                    .ok_or_else(|| {
+                        proto_error(format!(
+                            "Received an unknown scalar function: {}",
+                            expr.fun
+                        ))
+                    })?;
+                let args = &expr.args;
+
+                match scalar_function {
+                    ScalarFunction::Sqrt => Ok(sqrt((&args[0]).try_into()?)),
+                    ScalarFunction::Sin => Ok(sin((&args[0]).try_into()?)),
+                    ScalarFunction::Cos => Ok(cos((&args[0]).try_into()?)),
+                    ScalarFunction::Tan => Ok(tan((&args[0]).try_into()?)),
+                    // ScalarFunction::Asin => Ok(asin(&args[0]).try_into()?)),
+                    // ScalarFunction::Acos => Ok(acos(&args[0]).try_into()?)),
+                    ScalarFunction::Atan => Ok(atan((&args[0]).try_into()?)),
+                    ScalarFunction::Exp => Ok(exp((&args[0]).try_into()?)),
+                    ScalarFunction::Log2 => Ok(log2((&args[0]).try_into()?)),
+                    ScalarFunction::Ln => Ok(ln((&args[0]).try_into()?)),
+                    ScalarFunction::Log10 => Ok(log10((&args[0]).try_into()?)),
+                    ScalarFunction::Floor => Ok(floor((&args[0]).try_into()?)),
+                    ScalarFunction::Ceil => Ok(ceil((&args[0]).try_into()?)),
+                    ScalarFunction::Round => Ok(round((&args[0]).try_into()?)),
+                    ScalarFunction::Trunc => Ok(trunc((&args[0]).try_into()?)),
+                    ScalarFunction::Abs => Ok(abs((&args[0]).try_into()?)),
+                    ScalarFunction::Signum => Ok(signum((&args[0]).try_into()?)),
+                    ScalarFunction::Octetlength => {
+                        Ok(octet_length((&args[0]).try_into()?))
+                    }
+                    // // ScalarFunction::Concat => Ok(concat((&args[0]).try_into()?)),
+                    ScalarFunction::Lower => Ok(lower((&args[0]).try_into()?)),
+                    ScalarFunction::Upper => Ok(upper((&args[0]).try_into()?)),
+                    ScalarFunction::Trim => Ok(trim((&args[0]).try_into()?)),
+                    ScalarFunction::Ltrim => Ok(ltrim((&args[0]).try_into()?)),
+                    ScalarFunction::Rtrim => Ok(rtrim((&args[0]).try_into()?)),
+                    // ScalarFunction::Totimestamp => Ok(to_timestamp((&args[0]).try_into()?)),
+                    // ScalarFunction::Array => Ok(array((&args[0]).try_into()?)),
+                    // // ScalarFunction::Nullif => Ok(nulli((&args[0]).try_into()?)),
+                    ScalarFunction::Datepart => {
+                        Ok(date_part((&args[0]).try_into()?, (&args[1]).try_into()?))
+                    }
+                    ScalarFunction::Datetrunc => {
+                        Ok(date_trunc((&args[0]).try_into()?, (&args[1]).try_into()?))
+                    }
+                    // ScalarFunction::Md5 => Ok(md5((&args[0]).try_into()?)),
+                    ScalarFunction::Sha224 => Ok(sha224((&args[0]).try_into()?)),
+                    ScalarFunction::Sha256 => Ok(sha256((&args[0]).try_into()?)),
+                    ScalarFunction::Sha384 => Ok(sha384((&args[0]).try_into()?)),
+                    ScalarFunction::Sha512 => Ok(sha512((&args[0]).try_into()?)),
+                    ScalarFunction::Digest => {
+                        Ok(digest((&args[0]).try_into()?, (&args[1]).try_into()?))
+                    }
+                    _ => Err(proto_error(
+                        "Protobuf deserialization error: Unsupported scalar function",
+                    )),
+                }
+            }
+        }
+    }
+}
+
+impl TryFrom<protobuf::WindowFrame> for WindowFrame {
+    type Error = Error;
+
+    fn try_from(window: protobuf::WindowFrame) -> Result<Self, Self::Error> {
+        let units = protobuf::WindowFrameUnits::from_i32(window.window_frame_units)
+            .ok_or_else(|| {
+                proto_error(format!(
+                    "Received a WindowFrame message with unknown WindowFrameUnits {}",
+                    window.window_frame_units
+                ))
+            })?
+            .into();
+        let start_bound = window
+            .start_bound
+            .ok_or_else(|| {
+                proto_error(
+                    "Received a WindowFrame message with no start_bound".to_owned(),
+                )
+            })?
+            .try_into()?;
+        let end_bound = window
+            .end_bound
+            .map(|end_bound| match end_bound {
+                protobuf::window_frame::EndBound::Bound(end_bound) => {
+                    end_bound.try_into()
+                }
+            })
+            .transpose()?
+            .unwrap_or(WindowFrameBound::CurrentRow);
+        Ok(Self {
+            units,
+            start_bound,
+            end_bound,
+        })
+    }
+}
+
+impl TryFrom<protobuf::WindowFrameBound> for WindowFrameBound {
+    type Error = Error;
+
+    fn try_from(bound: protobuf::WindowFrameBound) -> Result<Self, Self::Error> {
+        let bound_type = protobuf::WindowFrameBoundType::from_i32(bound.window_frame_bound_type).ok_or_else(|| {
+            proto_error(format!(
+                "Received a WindowFrameBound message with unknown WindowFrameBoundType {}",
+                bound.window_frame_bound_type
+            ))
+        })?;
+        match bound_type {
+            protobuf::WindowFrameBoundType::CurrentRow => Ok(Self::CurrentRow),
+            protobuf::WindowFrameBoundType::Preceding => {
+                // FIXME implement bound value parsing
+                // https://github.com/apache/arrow-datafusion/issues/361
+                Ok(Self::Preceding(Some(1)))
+            }
+            protobuf::WindowFrameBoundType::Following => {
+                // FIXME implement bound value parsing
+                // https://github.com/apache/arrow-datafusion/issues/361
+                Ok(Self::Following(Some(1)))
+            }
+        }
+    }
+}
+
+impl TryFrom<&Box<protobuf::List>> for DataType {
+    type Error = Error;
+    fn try_from(list: &Box<protobuf::List>) -> Result<Self, Self::Error> {
+        let list_ref = list.as_ref();
+        match &list_ref.field_type {
+            Some(pb_field) => {
+                let pb_field_ref = pb_field.as_ref();
+                let arrow_field: Field = pb_field_ref.try_into()?;
+                Ok(Self::List(Box::new(arrow_field)))
+            }
+            None => Err(proto_error(
+                "List message missing required field 'field_type'",
+            )),
+        }
+    }
+}
+
+impl protobuf::TimeUnit {
+    pub fn from_arrow_time_unit(val: &TimeUnit) -> Self {
+        match val {
+            TimeUnit::Second => protobuf::TimeUnit::Second,
+            TimeUnit::Millisecond => protobuf::TimeUnit::TimeMillisecond,
+            TimeUnit::Microsecond => protobuf::TimeUnit::Microsecond,
+            TimeUnit::Nanosecond => protobuf::TimeUnit::Nanosecond,
+        }
+    }
+
+    pub fn from_i32_to_arrow(time_unit_i32: i32) -> Result<TimeUnit, Error> {
+        let pb_time_unit = protobuf::TimeUnit::from_i32(time_unit_i32);
+        match pb_time_unit {
+            Some(time_unit) => Ok(match time_unit {
+                protobuf::TimeUnit::Second => TimeUnit::Second,
+                protobuf::TimeUnit::TimeMillisecond => TimeUnit::Millisecond,
+                protobuf::TimeUnit::Microsecond => TimeUnit::Microsecond,
+                protobuf::TimeUnit::Nanosecond => TimeUnit::Nanosecond,
+            }),
+            None => Err(proto_error(
+                "Error converting i32 to TimeUnit: Passed invalid variant",
+            )),
+        }
+    }
+}
+
+impl protobuf::IntervalUnit {
+    pub fn from_arrow_interval_unit(interval_unit: &IntervalUnit) -> Self {
+        match interval_unit {
+            IntervalUnit::YearMonth => protobuf::IntervalUnit::YearMonth,
+            IntervalUnit::DayTime => protobuf::IntervalUnit::DayTime,
+            IntervalUnit::MonthDayNano => protobuf::IntervalUnit::MonthDayNano,
+        }
+    }
+
+    pub fn from_i32_to_arrow(interval_unit_i32: i32) -> Result<IntervalUnit, Error> {
+        let pb_interval_unit = protobuf::IntervalUnit::from_i32(interval_unit_i32);
+        match pb_interval_unit {
+            Some(interval_unit) => Ok(match interval_unit {
+                protobuf::IntervalUnit::YearMonth => IntervalUnit::YearMonth,
+                protobuf::IntervalUnit::DayTime => IntervalUnit::DayTime,
+                protobuf::IntervalUnit::MonthDayNano => IntervalUnit::MonthDayNano,
+            }),
+            None => Err(proto_error(
+                "Error converting i32 to DateUnit: Passed invalid variant",
+            )),
+        }
+    }
+}
+
+fn vec_to_array<T, const N: usize>(v: Vec<T>) -> [T; N] {
+    v.try_into().unwrap_or_else(|v: Vec<T>| {
+        panic!("Expected a Vec of length {} but it was {}", N, v.len())
+    })
+}
+
+//Does not typecheck lists
+fn typechecked_scalar_value_conversion(
+    tested_type: &protobuf::scalar_value::Value,
+    required_type: protobuf::PrimitiveScalarType,
+) -> Result<ScalarValue, Error> {
+    use protobuf::{scalar_value::Value, PrimitiveScalarType};
+
+    Ok(match (tested_type, &required_type) {
+        (Value::BoolValue(v), PrimitiveScalarType::Bool) => {
+            ScalarValue::Boolean(Some(*v))
+        }
+        (Value::Int8Value(v), PrimitiveScalarType::Int8) => {
+            ScalarValue::Int8(Some(*v as i8))
+        }
+        (Value::Int16Value(v), PrimitiveScalarType::Int16) => {
+            ScalarValue::Int16(Some(*v as i16))
+        }
+        (Value::Int32Value(v), PrimitiveScalarType::Int32) => {
+            ScalarValue::Int32(Some(*v))
+        }
+        (Value::Int64Value(v), PrimitiveScalarType::Int64) => {
+            ScalarValue::Int64(Some(*v))
+        }
+        (Value::Uint8Value(v), PrimitiveScalarType::Uint8) => {
+            ScalarValue::UInt8(Some(*v as u8))
+        }
+        (Value::Uint16Value(v), PrimitiveScalarType::Uint16) => {
+            ScalarValue::UInt16(Some(*v as u16))
+        }
+        (Value::Uint32Value(v), PrimitiveScalarType::Uint32) => {
+            ScalarValue::UInt32(Some(*v))
+        }
+        (Value::Uint64Value(v), PrimitiveScalarType::Uint64) => {
+            ScalarValue::UInt64(Some(*v))
+        }
+        (Value::Float32Value(v), PrimitiveScalarType::Float32) => {
+            ScalarValue::Float32(Some(*v))
+        }
+        (Value::Float64Value(v), PrimitiveScalarType::Float64) => {
+            ScalarValue::Float64(Some(*v))
+        }
+        (Value::Date32Value(v), PrimitiveScalarType::Date32) => {
+            ScalarValue::Date32(Some(*v))
+        }
+        (Value::TimeMicrosecondValue(v), PrimitiveScalarType::TimeMicrosecond) => {
+            ScalarValue::TimestampMicrosecond(Some(*v), None)
+        }
+        (Value::TimeNanosecondValue(v), PrimitiveScalarType::TimeMicrosecond) => {
+            ScalarValue::TimestampNanosecond(Some(*v), None)
+        }
+        (Value::Utf8Value(v), PrimitiveScalarType::Utf8) => {
+            ScalarValue::Utf8(Some(v.to_owned()))
+        }
+        (Value::LargeUtf8Value(v), PrimitiveScalarType::LargeUtf8) => {
+            ScalarValue::LargeUtf8(Some(v.to_owned()))
+        }
+
+        (Value::NullValue(i32_enum), required_scalar_type) => {
+            if *i32_enum == *required_scalar_type as i32 {
+                let pb_scalar_type = PrimitiveScalarType::from_i32(*i32_enum).ok_or_else(|| {
+                    Error::General(format!(
+                        "Invalid i32_enum={} when converting with PrimitiveScalarType::from_i32()",
+                        *i32_enum
+                    ))
+                })?;
+                let scalar_value: ScalarValue = match pb_scalar_type {
+                    PrimitiveScalarType::Bool => ScalarValue::Boolean(None),
+                    PrimitiveScalarType::Uint8 => ScalarValue::UInt8(None),
+                    PrimitiveScalarType::Int8 => ScalarValue::Int8(None),
+                    PrimitiveScalarType::Uint16 => ScalarValue::UInt16(None),
+                    PrimitiveScalarType::Int16 => ScalarValue::Int16(None),
+                    PrimitiveScalarType::Uint32 => ScalarValue::UInt32(None),
+                    PrimitiveScalarType::Int32 => ScalarValue::Int32(None),
+                    PrimitiveScalarType::Uint64 => ScalarValue::UInt64(None),
+                    PrimitiveScalarType::Int64 => ScalarValue::Int64(None),
+                    PrimitiveScalarType::Float32 => ScalarValue::Float32(None),
+                    PrimitiveScalarType::Float64 => ScalarValue::Float64(None),
+                    PrimitiveScalarType::Utf8 => ScalarValue::Utf8(None),
+                    PrimitiveScalarType::LargeUtf8 => ScalarValue::LargeUtf8(None),
+                    PrimitiveScalarType::Date32 => ScalarValue::Date32(None),
+                    PrimitiveScalarType::TimeMicrosecond => {
+                        ScalarValue::TimestampMicrosecond(None, None)
+                    }
+                    PrimitiveScalarType::TimeNanosecond => {
+                        ScalarValue::TimestampNanosecond(None, None)
+                    }
+                    PrimitiveScalarType::Null => {
+                        return Err(proto_error(
+                            "Untyped scalar null is not a valid scalar value",
+                        ))
+                    }
+                    PrimitiveScalarType::Decimal128 => {
+                        ScalarValue::Decimal128(None, 0, 0)
+                    }
+                    PrimitiveScalarType::Date64 => ScalarValue::Date64(None),
+                    PrimitiveScalarType::TimeSecond => {
+                        ScalarValue::TimestampSecond(None, None)
+                    }
+                    PrimitiveScalarType::TimeMillisecond => {
+                        ScalarValue::TimestampMillisecond(None, None)
+                    }
+                    PrimitiveScalarType::IntervalYearmonth => {
+                        ScalarValue::IntervalYearMonth(None)
+                    }
+                    PrimitiveScalarType::IntervalDaytime => {
+                        ScalarValue::IntervalDayTime(None)
+                    }
+                };
+                scalar_value
+            } else {
+                return Err(proto_error("Could not convert to the proper type"));
+            }
+        }
+        (Value::Decimal128Value(val), PrimitiveScalarType::Decimal128) => {
+            let array = vec_to_array(val.value.clone());
+            ScalarValue::Decimal128(
+                Some(i128::from_be_bytes(array)),
+                val.p as usize,
+                val.s as usize,
+            )
+        }
+        (Value::Date64Value(v), PrimitiveScalarType::Date64) => {
+            ScalarValue::Date64(Some(*v))
+        }
+        (Value::TimeSecondValue(v), PrimitiveScalarType::TimeSecond) => {
+            ScalarValue::TimestampSecond(Some(*v), None)
+        }
+        (Value::TimeMillisecondValue(v), PrimitiveScalarType::TimeMillisecond) => {
+            ScalarValue::TimestampMillisecond(Some(*v), None)
+        }
+        (Value::IntervalYearmonthValue(v), PrimitiveScalarType::IntervalYearmonth) => {
+            ScalarValue::IntervalYearMonth(Some(*v))
+        }
+        (Value::IntervalDaytimeValue(v), PrimitiveScalarType::IntervalDaytime) => {
+            ScalarValue::IntervalDayTime(Some(*v))
+        }
+        _ => return Err(proto_error("Could not convert to the proper type")),
+    })
+}
+
+fn from_proto_binary_op(op: &str) -> Result<Operator, Error> {
+    match op {
+        "And" => Ok(Operator::And),
+        "Or" => Ok(Operator::Or),
+        "Eq" => Ok(Operator::Eq),
+        "NotEq" => Ok(Operator::NotEq),
+        "LtEq" => Ok(Operator::LtEq),
+        "Lt" => Ok(Operator::Lt),
+        "Gt" => Ok(Operator::Gt),
+        "GtEq" => Ok(Operator::GtEq),
+        "Plus" => Ok(Operator::Plus),
+        "Minus" => Ok(Operator::Minus),
+        "Multiply" => Ok(Operator::Multiply),
+        "Divide" => Ok(Operator::Divide),
+        "Modulo" => Ok(Operator::Modulo),
+        "Like" => Ok(Operator::Like),
+        "NotLike" => Ok(Operator::NotLike),
+        other => Err(proto_error(format!(
+            "Unsupported binary operator '{:?}'",
+            other
+        ))),
+    }
+}
+
+fn parse_required_expr(
+    p: &Option<Box<protobuf::LogicalExprNode>>,
+) -> Result<Expr, Error> {
+    match p {
+        Some(expr) => expr.as_ref().try_into(),
+        None => Err(proto_error("Missing required expression")),
+    }
+}
+
+fn parse_optional_expr(
+    p: &Option<Box<protobuf::LogicalExprNode>>,
+) -> Result<Option<Expr>, Error> {
+    match p {
+        Some(expr) => expr.as_ref().try_into().map(Some),
+        None => Ok(None),
+    }
+}
+
+fn proto_error<S: Into<String>>(message: S) -> Error {
+    Error::General(message.into())
+}
